Added test for flag :skip_scm (which should be set in the main deploy.(rb|rake) file that if set, will skip anything to do with the scm directory on the remote host and in all actuality it will not even set up the scm directory if the :setup task is called.
The need for this arises when using subversion and if you follow the procedure of actually having a new tag made (in subversion) before deploying code to the remote host(s).  If in the initial deploy you push out the tag (http://some.server/some-project/tags/REL-1) then on a subsequent push try and release (http://some.server/some-project/tags/REL-2) vlad (due to a subversion error) will croak due to it attempting to perform an `svn up `in the scm directory.  The .svn files in the scm directory are pointed at REL-1 and subversion will throw an error somewhat along the lines of "can't switch repositories" when trying to deploy REL-2

When setting the :skip_scm flag, (set :skip_scm, true) it makes vlad skip the `svn update` all together and just runs `svn export` (which is does anyhow).

Summary -- Bypasses the issues with the scm directory and `svn update`